backend/app/database_api/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
pydantic
pydantic[email]
boto3
python-multipart